+++ src/download.rs
@@ -29,6 +29,7 @@ pub struct Asset {
 pub struct Release {
     pub tag_name: String,
     pub draft: bool,
+    pub body: Option<String>,
     pub assets: Vec<Asset>,
 }
 
@@ -356,6 +357,7 @@ mod tests {
         Release {
             tag_name: tag.to_owned(),
             draft: false,
+            body: None,
             assets: Vec::new(),
         }
     }